Added decPAcked
The "Packed" name is not set in stone; it may become BCDPacked.

Will be handy for constants. Also changed the parameter of NewNumber
from *Context to int32. This was motivated by need to be able to create
context free Number's (like large precision constants) and give the
ability to modules like Packed to create Number's.

decPacked, decQuad, decimal64 and decimal128 are now compiled in, ready
to be Go-y-fied.